+++ src/main/java/org/example/model/game/elements/towers/MageTower.java
@@ -4,7 +4,6 @@ package org.example.model.game.elements.towers;
 import org.example.model.game.elements.enemys.Enemy;
 
 import java.util.ArrayList;
-import java.util.List;
 
 public class MageTower extends Tower {
 
@@ -14,7 +13,7 @@ public class MageTower extends Tower {
     private static final int COST = 500;
     private final int firingSpeed,damage;
 
-    public MageTower(int x, int y){
+    public MageTower(float x, float y){
         super(MT_LIFE, LEVEL, RANGE, COST,x,y,new ArrayList<>());
         this.towerSymbol = 'M';
         this.damage=50;
